Wikimedia instance dead
I noticed that anon wasn't running for a couple months by accident. As I
tried to get it working again I noticed that the Kubernetes cluster and
NFS seemed to be malfunctioning and I didn't have the time to
investigate what was going on other than opening a ticket:

https://phabricator.wikimedia.org/T262734

For the moment I've taken the config.json from the Wikimedia instance
and am running it on my own server.